requirements.txt
PyQt6
librosa
websockets
numpy
opencolorio
